# demos/hypermedia-server/static/meta/openapi.json
{
  "info": {
    "title": "Hypermedia Demo API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/file-objects": {
      "get": {
        "operationId": "getFileObjects",
        "parameters": [
          {
            "description": "Opaque pagination cursor returned by a previous response",
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Maximum number of items to return",
            "in": "query",
            "name": "limit",
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Note identifier",
            "in": "query",
            "name": "noteId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "urn:schema:hal:storage.FileObjectCollection:1.0.0"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "urn:schema:storage.FileObjectCollection:1.0.0"
                }
              }
            },
            "description": "Success"
          }
        },
        "tags": ["FileObject"]
      },
      "post": {
        "operationId": "getFileUploadPermit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "urn:schema:storage.CreateFileObject:1.0.0"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "urn:schema:storage.PresignedPermitResponse:1.0.0"
                }
              }
            },
            "description": "Success"
          }
        },
        "tags": ["FileObject"]
      }
    },
    "/api/file-objects/{id}": {
      "get": {
        "operationId": "getFileObjectById",
        "parameters": [
          {
            "description": "File object identifier",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "urn:schema:hal:storage.FileObject:1.0.0"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "urn:schema:storage.FileObject:1.0.0"
                }
              }
            },
            "description": "Success"
          }
        },
        "tags": ["FileObject"]
      }
    },
...
  },
...
}